webapp/requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.2.1
distribute==0.6.24
itsdangerous==0.24
solrpy==0.9.7
wsgiref==0.1.2
